chore: use fetch instead of request
