+++ packages/core/src/redocly/redocly-client-types.ts
@@ -8,14 +8,3 @@ export interface RegionalToken {
 export interface RegionalTokenWithValidity extends RegionalToken {
   valid: boolean;
 }
-
-export type Analytics = {
-  event: string;
-  event_time: string;
-  logged_in: boolean;
-  command: string | number;
-  arguments: Record<string, unknown>;
-  node_version: string;
-  version: string;
-  exit_code: 0 | 1;
-};
